flask/requirements.txt
Flask
python-dotenv
gspread
google-auth
